--- a/package.json
+++ b/package.json
@@ -137,7 +137,7 @@
     "cross-env": "^7.0.3",
     "del-cli": "^6.0.0",
     "eslint": "^9.23.0",
-    "eslint-import-resolver-oxc": "^0.13.1",
+    "eslint-import-resolver-typescript": "^4.2.4",
     "eslint-plugin-import-x": "^4.9.1",
     "eslint-plugin-n": "^17.16.2",
     "eslint-plugin-sort-destructure-keys": "^2.0.0",