Hi, `eslint-plugin-import-x` and `eslint-import-resolver-typescript` owner here.

The brand new `eslint-import-resolver-typescript` is now powered by [unrs-resolver](https://github.com/unrs/unrs-resolver) which is fork of [oxc-resolver](https://github.com/oxc-project/oxc-resolver) and [rspack-resolver](https://github.com/web-infra-dev/rspack-resolver) with P'n'P support and more targets, and it has competitive performance compared with [`eslint-import-resolver-oxc`](https://github.com/9romise/eslint-import-resolver-oxc).

Before with `resolver-oxc`:

```sh
 socket-cli on  main [$!] ❯ TIMING=1 npmR check:lint

> socket@0.14.67 check:lint
> eslint --report-unused-disable-directives .

Rule                                | Time (ms) | Relative
:-----------------------------------|----------:|--------:
import-x/namespace                  |  1131.479 |    62.1%
@typescript-eslint/return-await     |   288.136 |    15.8%
import-x/order                      |    98.431 |     5.4%
import-x/extensions                 |    30.366 |     1.7%
n/no-extraneous-import              |    19.631 |     1.1%
n/no-unsupported-features/es-syntax |    18.692 |     1.0%
n/no-extraneous-require             |    17.510 |     1.0%
import-x/no-duplicates              |     9.148 |     0.5%
no-misleading-character-class       |     8.638 |     0.5%
import-x/default                    |     8.380 |     0.5%
is 📦 v0.14.67 via  v23.7.0 took 15s
```

After with `resolver-typescript`:

```sh
 socket-cli on  main [+] ❯ TIMING=1 npmR check:lint

> socket@0.14.67 check:lint
> eslint --report-unused-disable-directives .

Rule                                | Time (ms) | Relative
:-----------------------------------|----------:|--------:
import-x/namespace                  |  1058.428 |    65.0%
@typescript-eslint/return-await     |   225.935 |    13.9%
import-x/order                      |    71.322 |     4.4%
import-x/extensions                 |    27.978 |     1.7%
n/no-extraneous-require             |    15.662 |     1.0%
n/no-extraneous-import              |    15.299 |     0.9%
n/no-unpublished-import             |    12.542 |     0.8%
n/no-unsupported-features/es-syntax |    11.428 |     0.7%
n/no-unpublished-require            |     9.040 |     0.6%
no-misleading-character-class       |     8.839 |     0.5%
is 📦 v0.14.67 via  v23.7.0 took 10s
```