@JounQin 2025-03-26
Actually, rules like `n/no-extraneous-require`/`n/no-extraneous-import` can just be remove in favor of `no-extraneous-dependencies` alreay covering them.