@JounQin 2025-03-26
@jdalton By the way, I have troubles when running `npm run bs` locally

```sh
 socket-cli on  chore/import_resolver ❯ npmR bs

> socket@0.14.67 bs
> npm run build:dist; npm exec socket --


> socket@0.14.67 build:dist
> del-cli 'dist' && rollup -c .config/rollup.dist.config.mjs


/Users/JounQin/Workspaces/GitHub/socket-cli/src/cli.ts, /Users/JounQin/Workspaces/GitHub/socket-cli/src/constants.ts, /Users/JounQin/Workspaces/GitHub/socket-cli/src/shadow/npm/bin.ts, /Users/JounQin/Workspaces/GitHub/socket-cli/src/shadow/npm/inject.ts → dist/module-sync...
created dist/module-sync in 5.3s

/Users/JounQin/Workspaces/GitHub/socket-cli/src/cli.ts, /Users/JounQin/Workspaces/GitHub/socket-cli/src/constants.ts, /Users/JounQin/Workspaces/GitHub/socket-cli/src/shadow/npm/bin.ts, /Users/JounQin/Workspaces/GitHub/socket-cli/src/shadow/npm/inject.ts → dist/require...
created dist/require in 4.9s
node:internal/modules/cjs/loader:1397
  throw err;
  ^

{
  code: 'MODULE_NOT_FOUND',
  requireStack: [
    '/Users/JounQin/Workspaces/GitHub/socket-cli/node_modules/@pnpm/lockfile-file/lib/logger.js',
    '/Users/JounQin/Workspaces/GitHub/socket-cli/node_modules/@pnpm/lockfile-file/lib/write.js',
    '/Users/JounQin/Workspaces/GitHub/socket-cli/node_modules/@pnpm/lockfile-file/lib/index.js',
    '/Users/JounQin/Workspaces/GitHub/socket-cli/dist/module-sync/cli.js'
  ]
}

Node.js v23.7.0
```